BUG: Install required dependencies
This commit updates the build-system to use a "SuperBuild" structure
and install and required packages when built as a Slicer extension.

Note that the packages are not explicitly packaged when the extension
is bundled into a Slicer-based application. In this case, the variable
Slicer_ADDITIONAL_DEPENDENCIES may be updated by appending
"python-MFSDA-requirements".